Improved method for generating base tool-chain
The method for searching for isntalled cache files has received a
large overhaul, with borrowings from sorcery itself. The host triplet
is gathered from the sorcery archspec files corresponding to the value
of $ARCHITECTURE defined in /etc/local/sorcery/config. The version of
the spell to grab the cache tarball for is taken from the version
which is currently installed, by grabbing the last field in
/var/state/sorcery/packages. A function for generating a cache tarball
when one doesn't exist on the host for the corresponding spell still
needs to be implemented though.