bashrc: new prompt, branching, better printing
Added the new prompt system, which implements a help system as well.
Related to this, there are two user-level commands that were added,
help() and welcome(). help() prints the list of commands, while
welcome() prints the welcome screen from boot-up.

A new function was added to handle branching for cases beyond mere
linear progression. This may be deprecated later.

A new header() function outputs 3 newlines and before cat'ing the
requested file. This improves readability of the information displayed
to the user.