keymap: new script to choose keymap
This script will parse /usr/share/keymaps for all possible keymaps
installed on the system (iso) and present them to the user in a menu
interface. The user can navigate the menus solely with the use of
number keys, so this will help minimize problems where users cannot
select the correct keymap if they don't know how to type in qwerty
(the default).