smgl-chroot: new file, eases setup of chroot env
A simple script that helps automate the setting up of a proper chroot
environment for the installation process. By default it echo's what it
is doing, but it can also run silently when passed the '-q' option.
This will help to greatly simplify the setup procedures for e.g.
bootloaders and changing the root passwd.