steps/user-setup: fixed the info for the useradd command
Added some flags to the example useradd command to correct it so that
home dirs are created at /home instead of /home/users, and so that new
users are created with GID 1000. Users should be able to follow the
format of the example to modify it if they want behavior other than
the default represented in the example.