help/index: added more information for users
Added some more helpful information, as well as a little crucial
information. The help file now informs the user of the new ways to use
the help command (as in with possible arguments) so that users can see
the editors file for example. Also made the info about the back
command more explicit, so that users will not confuse it with e.g. the
back button of a web browser.