scripts/mkinitrd: Fixed the libscan function/usage
Both the definition and usage of libscan were broken. The definition
failed to properly select all of the libs that would be returned by
ldd. The usage failed to take into account the fact that the results
returned via ldd would be a list, and therefore would need multiple
iterations to copy each one. This is now properly dealt with.