steps/user-setup: fixed the useradd instructions
Useradd is really, really broken. In order to get the expected
(correct?) behavior with regards to the homedir to be created, it has
to be manually typed in, in its entirety. This is extremely dumb on
the part of useradd (note that adduser is not really any better).
Also, I added in the -g users stuff. If a user wants to use different
groups, they can modify the command, so having it is better so that
they will know what the option is.