base/etc/lilo.conf: new default lilo conf
I went ahead and created a really basic default conf file for lilo.
This will need to be added to (in the way of comment structures) to
help clarify things for users, and perhaps to provide better defaults
than what I have given here. At least this is a start though, and it
will work for a basic system (tested via qemu with the newly released
devel ISO).