libcauldron: defined code for cleaning iso
Defined the code for the cauldron_clean_iso_system function. This
de-lints the iso from unnecessary/unwanted files that were installed
from spells. The intended primary use is to remove things like gcc,
but we can't remove it via dispel because some of its installed libs
are still needed. However, it is functional enough to remove any
file/directory. Absolute paths should be given relative to the root of
a chroot of the iso since this is intended to run from inside a chroot
for security purposes.