libcauldron,liberror: new error handling
Added the liberror file to take care of all of the error handling for
cauldron. This currently only has error definitions (variables working
as \#defines or enums), but will later also include error-handling
functions as well. I also updated some of the code in libcauldron to
make use of it.