cauldron: added 'brew' command
The brew command is for brewing an ISO (making one). For now this is
just a wrapper around the libcauldron cauldron_build_iso function,
which pulls in values from the system cauldron config file and goes
through the necessary steps to create an entire iso. This will likely
get changed in the future to be more granular.