added HACKING file
The HACKING file explains how the different components of cauldron all
fit together. It specifies where things get sourced, and what their
purposes are. This should make it easier for others to get acquainted
with the code, so that we can have more people helping out, and should
also improve the maintainability of the code.