TODO: list of things still needed to be done
This lists the immediate needs in order to get cauldron working enough
to produce functional ISOs (assuming everything already works of what
is there). It also is a place-holder for ideas of things that could be
implemented later (even if they never happen). The later things are of
course subject to change, and are just ideas (exception: no known bugs
is definitely wanted).