libcauldron: cauldron_build_iso adjustments
Reworked cauldron_build_iso to use the newer versions of other
functions whose arguments had changed. Also, corrected the use of
cauldron_add_installer so that it can add all the installers it is
supposed to instead of only a single installer. Currently installers
are to be specified in TARGET_SPELLS, but since they are not spells
this is kind of wrong and should change.