libinit: filled in all the functions
The majority of the functions assume the existence of files not yet
created in the CAULDRON_DATA directory. Also, they make the assumption
that all the bin locations will be symlinked to /bin on the initfs, so
all binaries can be copied to the same place. This may change in the
future.