back->prev, release_version
The back function was renamed to prev, for 2 reasons: 1) it was more
confusing to users since it didn't match with next, and 2) it will
make things easier if/when the back and forward browsing history is
pushed into mainline (currently in devel-flux branch).

Somehow @release@ got replaced to devel-20080207 on a previous commit
of mine. This fixes it back.