isospells.x86: removed raidtools from the list
The raidtools package has been deprecated in favor of mdadm for some
time. It's also failing to build, so I'm dropping it from the list of
basic spells to be on the ISO. It can be added as a special case, but
this list is for the basic spells that are required for all x86
release ISOs.