cleaniso.sh: ISO cleaning script
cleaniso.sh prunes the ISO filesystem of things it doesn't need. This
includes the build environment (gcc, g++, etc.). However, gcc and g++
both provide libraries that some binaries rely on during runtime, so a
simple "dispel $SPELL" would be overly destructive. The cleaners are
simple text listings of files/directories to be removed, one per line.
The format is exactly the same as the output from gaze install $SPELL,
except that the lines pertaining to files that need to remain on the
ISO are removed so that they won't be cleaned.