smgl.init: improved boot media selection algo
Improved the algorithms for how the initrd selects the media to boot
from. It will now use /sys instead of /proc to autodetect removable
media and try those for booting. I also added the missing boot-from
function so that users can now properly override the boot media
autodetection and force the initrd to try booting from a
user-specified device.