smgl.init: make udev handling cleaner
Don't kill the udev daemon until after the user has finished loading
modules and we are ready to pivot_root. Also, don't print FAILED to
STDOUT, since we can reasonably expect some modules to fail (we only
provide the essential modules on the initrd, so modules for the rest
of the kernel will fail). Maybe it will be better to only trigger for
certain subsystems, but this can be determined later.