steps/edit-confs: fixed "next" instruction
Erroneously said that the next step was bootloader, while it really is
the compile kernel step. Fixed by making it a generic "type next to go
to the next step" instruction, as the rest of the instructions are
(except for the bootloader step, where next goes to grub and a jump is
needed to get to lilo). Fixes bug #14787.