scripts/mkiso.sh: don't force smgl- in iso name
Made it so that the entire ISO name is configured by the first arg
($1) instead of forcibly prefixing the version with "smgl-". This will
make it easier to make non-release ISOs with custom names, since
mkrelease.sh can handle prepending "smgl-" to the version before
passing the string to mkiso.sh for official releases.