scripts/add-sauce.sh: use getopts to parse opts
Don't use real arguments to parse options, instead use getopts. This
lets users combine options (-fi for example) with no problems. Also,
we can combine the usage printing so that if wrong options are passed
or if the wrong number of options/arguments are passed we can print
the same usage easily.