scripts/add-sauce.sh: fix cp, add 'a' response
The cp algorithm now correctly copies. There were two problems
previously: the script failed to ensure that directories were created
if they were missing (shouldn't be a problem in theory as the
directories should exist, but now the script is better because it will
guarantee they exist) and it was copying directly into the base of
CHROOTDIR instead of the tree it belonged in. The 'a' response was
added as an option to "force" overwriting of files, so that if it's
your first time running add-sauce on a dir you don't have to type 'y'
on every file that matches.