scripts/cleaniso.sh: better error handling
Moved the usage info out into a real function so it can be used
multiple times. Added checks to make sure that the script is being
called correctly, as well as to make sure that the script is operating
on a sane chroot dir (make sure it has rm and rmdir).