implemented optional spell installation
Optional spells can now be installed in the install by calling
smgl-spellcaster. This script uses sorcery caches to install spells.
It presents a menu of all the caches it finds on the ISO, and the user
can toggle caches to be installed. Then it installs from the selected
list. It also remembers what spells were selected during the install
session (via the /tmp/spells file). This same mechanism allows us to
ship an ISO with a default selection of spells. Currently the script
makes a lot of assumptions and isn't that clever, but it will be
improved on in later versions.