scripts/spellcaster.sh: correctly cast in chroot
Modified the spellcaster script to correctly handle casting inside the
chroot. This requires running the script inside of the chroot by copying
it to the target directory along with the spell listings, and cleaning
up the target directory after we're done. This change makes use of the
fault checking now provided in cauldronchr.sh.