scripts/spellcaster.sh: separate cast and dispel
Casting and dispelling the ISO spells are now separate actions.
spellcaster.sh defaults to casting if no option is specified. This
separation was made because the ISO will need the optional spells to be
installed, while system.tar.bz2 should have only the required spells and
not the optional spells. Having the cast and dispel options separate
enables having all the spells present on the ISO, and the build dir can
just be copied and have the dispel run for the system.tar.bz2.