ChangeLog: removed - git handles version tracking
This file is not necessary, as all the history and version information
is handled by git. If this file is ever deemed necessary later on, it
can be pulled back out of the git history. All developers who modify
anything in the cauldron repository are instead required to write good
commit messgaes.