cauldron,libcauldron: refactor iso building
Replaced the "cauldron_build_iso" call in cauldron with the real code
for that function from libcauldron, and removed the code from
libcauldron. The function in libcauldron was just a convenience wrapper
around the other libcauldron functions, and didn't belong in libcauldron
since it was an end-product function. It now resides in its rightful
place in cauldron. This also allows for more flexibility in how the ISOs
will be generated since it is in cauldron.