libcauldron: quote vars, cond. rm cache files
Quoted path variables to ensure they won't be broken if there is a space
or some other strange character in the path elements. Also checked the
return code from untarring the cache files, and if it went ok the cache
files are removed from the base of the CAULDRON_BUILD dir (don't pollute
the build directory).