libcauldron: renamed iso_system to target_system
Renamed the function cauldron_build_iso_system to
cauldron_build_target_system. This is a much clearer name, since the
function is used to generate the system to be installed onto the target
rather than anything used on the ISO itself (i.e., it generates
the caches to be installed onto the target system).