data/ospells.x86: renamed to ispells.x86
The listing in ospells is really all of the spells that are in addition
to the system.tar.bz2 spells, and this includes what is needed for the
ISO to function as well as real "optional" spells. Therefore, this was
renamed to more accurately reflect the fact that these spells are needed
on the ISO as opposed to the system.