Properly separate spell listings
Updated the spell listing files to really be separate. This allows for
cleaner setups of the individual components. As a result, the only
interdependency is that tspells expects the caches of spells in ispells
to also be present, so the list of all available caches on a given ISO
will be caches(ispells) + caches(tspells).