scripts/spellcaster.sh: big overhaul
This script got a big overhaul. Most of the functionality the script was
providing has put into proper functions. This makes the code more
separable and easier to read. The script also supports optional spell
installation methods. Also, the script will no longer attempt to set up
everything "in-house", but will rather create spell caches in the build
directory and then copy/unpack those caches to separate directories
(system and iso) accordingly. This is a much cleaner method of
generation.