smgl.install/enchant: new, cache installer script
New script to support the installation of optional spell caches. The
script supports installing via /tmp/opt-spells, which can be set up by
alchemy (yet to be committed), or by handling caches one-off (which can
be looped to handle directories of caches). This is a powerful new
edition to the ISO installer, because not only does this allow for the
installation of optional spell caches provided on the ISO, it also
allows for one to install any caches from any medium that the ISO can
reach (and with net modules and net tools on the ISO, it's possible to
link up with nfs partitions).