scripts/add-sauce.sh: kernel is only for iso
The kernel only goes on the ISO and not in the sys dir, so don't
blacklist things in sys (there's no modules there to blacklist). Fixed
the CHROOTDIR variable to ensure that it doesn't have a trailing '/'.
Also, whether in sys or iso, run ldconfig afterwards, since presumably
etc/ld.so.conf will be copied over but the changes won't take effect if
the command isn't run via chroot.