scripts/spellcaster.sh: kernel version searching
The means of (possibly) obtaining the kernel version used in the build
have been improved. The kernel version can be passed as an optional
argument to the function. The function will scan the lib/modules
directory for different kernel versions before attemtping to get the
kernel version from the kernel config.