scripts/spellcaster.sh: improved setup_sys()
The sorcery and grimoire downloads now use a common dir on the TARGET
(TARGET/tmp). This will make it easier to clean up the files later. The
prefix path was missing from the untarring of the grimoire. Also, the
depends and packages files are force removed from SYSDIR so that we can
ensure we aren't writing duplicate entries in case the script is run
more than once with the same SYSDIR in existence.