scripts/mkinitrd.sh: better sanity checking
If the chroot dir (ISO_DIR) is given relative to $CWD, expand to an
absolute pathname. Quote uses of ISO_DIR in case spaces or other chars
are included in it. In addition to making sure the ISO_DIR kernel
modules can be found, also ensure that the ISO_DIR isolinux.cfg file can
be found.