scripts/mkrelease.sh: better grimoire version code
The real problem with the grimoire version sed string before was
actually not with the '-' being in the string, but with the fact that
the grimoire version info is actually two lines. The code was changed so
that GRIMOIRE_VER only holds the first line, which is the actual version
information, and doesn't include the second line which is the checksum.
The output is now correct.