scripts/spellcaster.sh: inc verbose output
Increased the verbosity of the script by adding a new function, msg,
that outputs it's argument if QUIET is not set. Correspondingly, there
is a new option for setting QUIET. Various messages were added using the
new msg function to make it clear what the script is doing as it's
processing things.