scripts/mkinitrd.sh: try to find kernel version
Try to automatically find the kernel version if it isn't specified on
the command-line. This works by checking to see if there is only one
kernel module directory in the ISO chroot, and using the version it
finds there. This should always work, because there should only be one
kernel version per ISO, but it's possible someone will use the scripts
to make an ISO with multiple kernel versions, in which case the version
will have to be specified as the first argument of the script.