smgl.install/smgl-enchant: better tablet creation
Don't pre-strip the gathered spell name before looping on it. This
prevents the script from accidentally stripping too much off. Also do
better testing to ensure that we have good values for generating
packages and depends (to be specific, ensure that the calls to
tablet_get_* succeeded as well as there actually being a spell-specific
depends file for information to write to depends).