scripts/spellcaster.sh: more build-only sanity
Added logic to sanity_check for build-only material, so that spellcaster
won't sanity check the build tree when not building. Also does better
build sanity checking by completely wiping the build tree etc and
setting it up for base defaults via add-sauce, and preparation of the
build sorcery is done as part of prepare_target.