scripts/spellcaster.sh: rm resolv.conf, due to ppp
Remove resolv.conf before trying to copy over it, since ppp will install
a symlink. This should become unnecessary once/if ppp can be fixed not
to force a symlink (or rather, fix it to install etc/ppp/resolv.conf as
a symlink to etc/resolv.conf, instead of the other way around like it
currently does).