scripts/qa.sh: new, does qaulity assurance
New script that does basic quality assurance on a build tree by checking
to see if any init scripts available for a spell are not installed, and
then offers to recast the spell. The user still needs to determine if
the offending spell really needs to be recast, since it will not always
be the case that we will want _all_ init scripts for a spell installed,
but this will help to identify cases where init scripts are missing.