liberror: use preserved return value
Preserve the return value to check before doing anything else in the
function. All tests, commands, etc. run will modify the value of $?, so
it needs to be preserved or what is tested later will actually be the
return value of function-internal processing, rather than the value the
caller wants to check.