liberror: correct and improve checkerror()
Corrected the previous faults in the implementation of checkerror(), as
well as improving it by adding additional functionality. It now makes
use of the liberror context variable and calls geterror() to get an
error name if one wasn't supplied. This enables checkerror() to be used
in a more generic way, in addition to only checking for a specific
error.