iso/etc/sysconfig/modutils: stop ISO boot depmod
Prevent the ISO from running depmod on bootup. First of all, it's not
needed since the kernel modules on the ISO are static (once they are on
a given ISO, they won't change). Second of all, the ISO is a read-only
filesystem so attempting to run depmod on it (which tries to write to
/lib/modules/`uname -r`/) causes a boot-time error.