new implmentation of liberror contexts
Rather than harcode the notion of calling contexts into liberror,
extract all the error codes and messages out to a separate file. This
file provides only base, default error codes and messages. Error
definitions specific to cauldron and enchantment are defined in their
respective source trees as errorcodes. This is a much more flexible
system, and also more optimized in terms of the code itself, as there is
less waste of variables and conditional branching.