automate copyright updates
Replace the hardcoded end year of our copyright with a placeholder
string in isolinux.msg, and use mkrelease.sh to replace that string with
the current year from the machine generating the ISO, at the time the
ISO is generated. This alleviates the need for ISO creators to remember
to update the copyright year manually (as long as the date/time of the
machine generating the ISO is correct).