scripts/spellcaster.sh: skip baseroot kernel check
When using spellcaster.sh to generate a new basesystem chroot, skip the
kernel checks and don't add any scripting for casting non-system spells
(i.e., we only want rspells in the chroot).

This work has not yet been tested, but should work. This commit can be
reverted pending testing that shows it doesn't do what it's supposed to.