spellcaster: forced optional depends in build
When spells are cast in the chroot, sorcery relies depends information
to determine the default for whether an optional dependency should be
enabled for the spell being cast. This commit implements a method for
forcing such dependencies as either enabled or disabled via listing the
spell, optional dependency, and enabled state  ("on" or "off")in
data/sorcery.depends. The contents of this file are used to overwrite
var/state/sorcery/depends in the build chroot before performing any
casting.