data/ispells.x86: add wpa_supplicant
Add wpa_supplicant, which may not necessarily be an absolute requirement
for getting wireless networking up and running (even with WPA; it
depends on the driver's features and one's familiarity manually
configuring for an ssid with wireless-tools), but definitely makes life
easier. It is provided in ispells rather than ospells for now because,
like the other base networking spells, although we don't advocate
connecting to net from inside the ISO runtime rather than an installed
system, we allow for it in case the user needs it and knows what he/she
is doing.