generalize the install "steps" to modules
Refer to install "steps" as modules since a single module may comprise
more than one step, or it may even be something other than a step in the
future. Also, don't tie modules to particular installers (the steps
file). Instead, use a single, separate directory to store all the
modules, and iterate over the modules directory to get the "steps" and
corresponding information.