add (reformatted) modules from the 0.10.0 series
All relevant "steps" from the 0.10.0 series are now present in devel.
They have been reformatted so that bold commands are marked with
@[ ... ]. In the future, additional markups may be implemented to
separate titles from commands, etc. The markup is to be replaced at
runtime (not install-time) by the installers. The modules were also
arranged so that help and welcome messages have their own directories,
since they are logically separate (but still modules).