liberror: reimplement error functions
The previous instantiation of liberror wasn't very useful. Error
checking should be the job of the caller, not liberror itself. The
present implementation of the functions is much more useful: error
messages that correspond to an error number are automatically looked up
and then printed in a specified color using libcolor functions; fatal
errors can be checked for and automatically cause exits after printing
the error in color.