enchantment: potion (state tracking) functionality
This is the initial implementation of potion, which is responsible for
state tracking in the install. This has 2 purposes: better
error-proofing the install itself, and later the state information can
be used for automated installs. When linked with the history
functionality already present in libenchantment though, we get a 3rd
added bonus essentially for free: (partial) undo in the installer, up to
as many history levels as configured in etc/enchantment/config in the
ISO (run-time configurable as well).