delint: new delinting script global_variables
We now have a script for detecting all global variables that are used in
given files and testing to see if they were defined in a given
definitions file (etc/config, errorcodes). This will help to ensure that
all global variables are properly defined before we get to run-time
testing, which will cut down on the time needed for debugging.