enchantment: rename color enabled function
Rename the color checking function to enchant_color_enabled to make its
name more straightforward for what it does. Now it should be more clear
that the function does not actually perform coloring, but rather only
checks to determine if color should be enabled/loaded.

Depending on how error checking is handled later, the functionality to
actually do the loading of lib.color might get incorporated into this
function if color should be enabled. For now it's handled in calling
functions/scripts, since the caller determines the severity of failing
to load lib.color when color is not expressly prohibited.