delint: better fault checking
Check to make sure that we have a good definition file before attempting
to match against it. If a check file doesn't exist though, just skip it
without emitting an error. We're only interested in catching undefined
variables, and there won't be any undefined variables to check for in a
file that isn't there.