update usage of libcolor functions in sources
Update all the uses of libcolor in cauldron and enchantment for the
recent changes in libcolor. Most importantly, make libcolor a
non-optional load, determining the optionality of colored output using
LIBCOLOR_NOCOLOR instead, and change uses of libcolor_message to
libcolor_notice. Also did a little bit of minor cleanup.