enchantment: new lib.i18n
Implemented the i18n sub-library. Currently it only handles
selecting/setting the keymap, but in the future it will also handle
which language the installer is currently running in, loading or
configuring input engines (if necessary), and anything else which may
need to be done in order to make the installer fully i18n capable.