enchantment: reimplement shell installer
Pushed functionality out of bashrc and into a separate and unified
installer command. All shell installer commands are now to be
implemented in this one script (enchantment), with core functionality
defined in the functions file (loaded by the enchantment command
script). I believe everything is implemented, but not all of it has been
tested. The potion functionality, however, is not yet implemented.