lib.sorcerous: API compliance
Rename functions to bring them in sync with the API specifications, and
make the sorcery download url a local variable (for now) so that it
stands out more in case of needed future changes. At some point it might
make sense to move it to a global configuration variable, but not for
now.