cauldron: explicitly deny chroot nesting
Force a return with error code if we are already in a chroot and
cauldron_chroot_init is called so that we avoid the mess of dealing with
nested chroots.

Nested chroots in and of themselves aren't bad, but we would have to
guarantee that we don't mount --bind multiple times to the same chroot
or cleaning up will get quite difficult. Tracking that would require at
least a stack for the chroots and their order, and some kind of testing
to ensure that the duplicates cannot occur. This may be implemented at
a later date, but for now it's better to keep the implementation simple.