cauldron: more logical functions in lib.sorcerous
The install function in lib.sorcerous previous did a lot more work than
it should have been doing (including download functions), while the
download function did almost nothing on its own.

This commit separates the functionality better, so that the install
function handles all and only install functionality, and the same goes
for the download function. It also moves some of the definitions out to
the config file so that there is better flexibility in how the
download and install are performed.