cauldron: remove superfluous libcauldron function
The function to build the system should be handled by a combination of
the functionality provided by lib.toolchain and commands processed by
the front-end. It shouldn't exist as a separate function in libcauldron.
The needed functionality is already present in lib.toolchain, so the
rest is up to the front-end (to appear), and the function can safely be
removed from libcauldron.