cauldron: generalize system/iso packaging
Rather than directly relying specifically on mkisofs, allow the user to
redefine the command to be used to whatever their preference is. Also
define a separate function for packaging the system (system.tar.bz2) and
generalize that as well to allow for different archiving/compression
schemes. Both functions now have clear names that follow the API
specification as well.