ward: make path management more generalized
Don't force path management to be specific to cauldron, and try not to
have hardcoded paths in general. This enables libward to be much more
flexible in handling file arguments so that it can be adapted for other
projects more easily. Also fixed a few leftover instances of
DELINT_PATH->WARD_PATH renamings I missed.