enchantment: shell: keymap can take 1 arg
The keymap command can either be called with no arguments, in which case
it will present a menu to the user from which they can select their
desired keymap. If called with an actual keymap name (any name
recognized directly by loadkeys, with or without path information), then
it will not present a menu and instead just set the keymap directly.