enchantment: potion: check->update
Change the lib.potion function names from *check* to *update*. The
functions perform checks to see if updates are needed, but they also
always perform the updates if available, so simply calling them "check"
is an inadequate description. The error code associated with the main
update function is also updated accordingly.