cauldron: add PROCESS to describe build process
Add a PROCESS file which describes the general cauldron build process
for both the sys and iso targets. This information could be helpful to
any new contributors (or anyone just curious). The entire process is
broken down into a list of sub-processes that should be separate (though
some are of course interdependent).