ward: handle sourced includes for config vars
Implemented a recursive descent parser to scan for sourced config files.
With the newly implemented cascading config file system in cauldron,
ward-configvars was unable to properly detect all the config variables
defined across the various files. The two options to fix this were to
maintain a list of all the possible config files, or to write a parser
that would scan all the files (otherwise, it would have to source the
files itself, which could be dangerous if the files contained actual
code). I opted for the parser, because this would additionally ensure
correctness by acting only on the files actually listed for inclusion,
and avoids the pitfall of having a source list that is out of sync with
the actual inclusions.