cauldron: convert all commands to use variables
All commands used in the cauldron sources are now externally
configurable via data/config/commands.conf, which is sourced by the
default profile config default.conf (which in turn is sourced by the
default user config file which would typically be installed to
/etc/cauldron/cauldron.conf). This makes cauldron much more configurable
for end users that would like to hack some advanced configuration,
and also makes cauldron more portable to systems that may not be using
the standard linux utilities (i.e., one of the BSDs, or a highly custom
system).

There is still some dependence in a few locations on the commands
providing certain output or expecting certain input, and there is one
instance where CAULDRON_CMD_TAR has to be used with its arguments
specified directly in the source file, making the use of tar less robust
(this is currently unavoidable, but I'll try and find a way around it if
I can). It should still be easy to find with a simple grep though, as
there is only one such instance (in lib.toolchain). There is less that
can be done about the assumptions regarding input and output without
making the configuration system useless, though. If inputs and outputs
really need to change, it would probably be better to directly edit the
sources themselves to port them.