cauldron: arch-specific build data/configs
Files that are necessary for listing the spells, sources, and caches
that will populate the iso and sys targets, the spells necessary for
minimal build toolchains, and configurations and listings needed for
generating initramfs/initramdisk targets. These may need to be adjusted
in the future. In particular, the initfs files are currently incomplete
(at present only the busybox config is present, which may need to be
tuned/altered later).