ward: refactor as a more general test harness
Remove Cauldron-specific hardcoding, and make it a more general test
harness. Currently the refactoring only applies to libward.

User-level commands are still Cauldron-specific, but perhaps eventually
libward, libruntime, and libsource (as well as liberror and libcolor)
will be moved out into their own repository if other components also
start making use of them, in which case leaving the user-level commands
as is will be fine since they are in the Cauldron repository.