ward: fix path variable setting/naming in userland
Correct the WARD_PATH setting in the generic user-level commands
functions and global_variables.

Although all the ward-* commands are Cauldron-specific, these two
commands are not, and should be usable as-is. However, both functions
and global_variables depend on the notion of a defined API for
function/variable naming. The variable naming matches the general
practice for shell global variables (all capitals, with numbers
and underscores allowed), though is more specific in that numbers are
only allowed as suffixes. The function naming presumes that "library"
file names consist of the general name (the function "prefix", e.g.
cauldron_) preceded by "lib" (e.g. libcauldron). This isn't hard to
modify, and if there is interest I could even refactor it to offload the
naming pattern into a separate file or variable.