lib/liberror: fold count into liberror_add_error
Currently, the implementation does not allow/account for either removing
error codes/messages, or adding them spuriously. They are always only
appended to the end of the list. There is therefore no need for
maintaining a separate count variable of the number of defined
codes/messages, as the same value can be ascertained simply by taking
the current size of the message array. This assumes that there are
always messages matching codes and vice versa, but this is guaranteed by
the use of liberror_add_error (if you manually adjust the codes or
messages, then it is your responsibility to account for any problems).
Folding the count tracking into liberror_add_error makes the
implementation a little cleaner and more localized.