lib/liberror: don't force stdout/stderr
Only define LIBERR_STDOUT and LIBERR_STDERR if they aren't already
defined. This allows for more flexibility in redirecting error message
output, say to a log file. Normally this could be overridden later, but
being able to redirect earlier on is more essential, since there may be
errors that occur before an override is to be performed in the loading
chain.