lib/liberror: allow for additional error info
Enable liberror_print_error to output additional error message
information after the regular error message. This can be used, for
example, to report via the regular error message that there was a
problem creating a file, and then through the additional information
which file failed to be created. liberror_die was also updated to allow
for passing this information to liberror_print_error.