lib/liberror: error checking functions
Added a new function, liberror_check, to test for the presence of a
non-fatal error on the error stack ($?). If it finds such an error, a
corresponding error message is printed via liberror_print_error. Also
rewrote liberror_check_fatal_error to leverage the newer liberror_die
(and in turn liberror_print_error) better, and renamed it to
liberror_check_fatal (less typing is better). Both functions can
optionally pass extended error message information along, as well as
optionally override the error code used from the one on the stack to one
supplied directly as the first parameter to the function.