don't hardcode /dev/null
Allow the _NULL variables (CAULDRON and ENCHANT) to point to paths other
than /dev/null, and only set the value to /dev/null if the _NULL
variables aren't already defined. This allows for overriding the
throwing away of what is normally considered unncessary output, so that
it can, for example, be collected in a debug log file. Also, ensure the
variables are properly quoted, as arbitrary paths may contain things
like spaces.