force "C" collating
Different locale settings can interfere with, for example, sorting or
grepping. Enforce sane locale settings for our purposes. This should not
affect end-user output from the scripts. If there is a problem in the
end-user output, then the scripts can be modified to first save the
locale settings and then restore them before final output.