enchantment: libenchantment: return real errorcode
When returning an error code after attempting libpotion_update, return
whatever error code is actually present rather than trying to force it
to a static one. The libpotion_update function is just a wrapper that
calls the collection of update functions, but if one of those functions
fails, libpotion_update will stop processing and immediately return an
error. Thus, if we return $? in libenchantment, we are getting the error
code of the function that really errored.