cauldron: toolchain learns to use baseroots
lib.toolchain has gained the ability to use pre-built basesystem
chroots. The baseroot can either be downloaded fresh, unpacked from an
already downloaded/created (compressed) archive, or an already unpacked
directory (i.e., a chroot directory ready for use). There are also
functions to verify the usability of the chroot post-unpacking, by
ensuring that it has some necessary directories as well as being able to
compile and execute a very simple C source file.

For later cross-compilation functionality (an eventual goal of cauldron
is automated cross-compilation to be able to build farm ISOs for
multiple target architectures), the compiler test will have to be
modified to not attempt to run the compiled binary, unless through a VM,
if the chroot is to be used for cross builds.